docker_mqtt/backend/ml_hub/logistic_regression/requirements.txt
pandas==1.0.1
scikit_learn==0.23.2
xlrd==1.2.0
paho-mqtt